pyproject.toml
[tool.poetry]
name = "chatstero"
version = "0.3.6"
description = "A simple markov chatbot."
authors = ["Kaylynn <mkaylynn7@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.5.0"
"markovify" = "^1.5.0"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
